monorepo/packages/ml_server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ml-server"
version = "0.1.0"
description = "ML_SERVER"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "uvicorn[standard]>=0.37.0",
    "shared",
]


[project.optional-dependencies]

# ML 공통 베이스 (torch + 기본 라이브러리)
ml-base = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "opencv-python>=4.9.0",
    "Pillow>=10.2.0",
    "numpy>=2.0.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.7.2",
    "scikit-image>=0.25.0",
]

# 공통 OCR 의존성
ocr-base = [
    "python-Levenshtein>=0.23.0",
]

# OCR CPU 버전 (경량화)
ocr-cpu = [
    "paddleocr>=2.8.0,<3.0.0",
    "paddlepaddle>=2.6.2,<3.0.0",  # CPU 버전 (GPU 버전보다 훨씬 가벼움)
    "easyocr>=1.7.2",
    "ml-server[ml-base]",
    "ml-server[ocr-base]",
]

# OCR GPU 버전
ocr-gpu = [
    "paddleocr>=2.8.0,<3.0.0",
    "paddlepaddle-gpu>=2.6.2,<3.0.0",  # GPU 버전
    "easyocr>=1.7.2",
    "ml-server[ml-base]",
    "ml-server[ocr-base]",
]

# 하위 호환성 유지 (기본은 GPU)
ocr = [
    "ml-server[ocr-gpu]",
]


[tool.hatch.build.targets.wheel]
packages = ["app"]